requirements.txt
python-dotenv==1.1.0
fastapi>=0.115
uvicorn>=0.34
